+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-import store from './store'
+import store from './reducers/store'
 import { Provider } from 'react-redux'
 import './index.css'
