+++ src/reducers/courseReducer.js
@@ -1,50 +1,20 @@
-import courseService from '../services/courses'
 
-const courseReducer = (store = [], action) => {
+// returns courses from store
+const courseReducer = (state = [], action) => {
   switch (action.type) {
   case 'CREATE_COURSE':
-    if (store.find(course => course.course_id === action.data.course_id)) {
-      return store
+    if (state.find(course => course.course_id === action.data.course_id)) {
+      return state
     }
-    return [...store, action.data]
+    return [...state, action.data]
 
   case 'INIT_COURSES':
     return action.data
 
   default:
-    return store
+    return state
   }
 }
 
-// Action creators
-
-export const initializeCourses = () => {
-  return async (dispatch) => {
-    const content = await courseService.getAll()
-    dispatch({
-      type: 'INIT_COURSES',
-      data: content
-    })
-  }
-}
-
-export const createContent = (content) => {
-  return async (dispatch) => {
-
-    const response = await courseService.create(content)
-
-    dispatch({
-      type: 'CREATE_COURSE',
-      data: response.course
-
-    })
-
-    dispatch({
-      type: 'CREATE_STUDENT',
-      data: response.student
-    })
-  }
-
-}
 
 export default courseReducer
